backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
onnxruntime==1.19.2
lightgbm==4.5.0
optuna==3.6.1
opencv-python-headless==4.10.0.84
joblib==1.4.2
shap==0.46.0
matplotlib==3.9.2
PyYAML==6.0.2
pyarrow==17.0.0
torch==2.1.0
torchvision==0.17.0
Pillow==10.0.0
